refract and move respective private methods to document-generator.util file and fix specs accordingly
