+++ test/spec/documentation/documentation.get.spec.ts
@@ -6,6 +6,7 @@ import {
   DocumentationService,
   ErrorReadingFile,
 } from '../../../src/apps/backend/modules/documentation';
+import DocumentGeneratorUtil from '../../../src/apps/backend/modules/documentation/internals/document-generator-util';
 import { OpenAIService } from '../../../src/apps/backend/modules/openai';
 import { app } from '../../helpers/app';
 
@@ -26,17 +27,6 @@ describe('GET /api/get-documentation', () => {
     sinonSandbox.restore();
   });
 
-  it('should return an empty string, as the documentation is disabled for the testing environment', async () => {
-    const res = await chai
-      .request(app)
-      .get('/api/get-documentation')
-      .set('content-type', 'application/json');
-
-    expect(res.status).to.be.eq(200);
-    expect(openAiStub.calledOnce).to.be.false;
-    expect(res.body.markdownDocumentation).to.be.eq('');
-  });
-
   it('should return the documentation when documentation is enabled for the environment', async () => {
     // stub the config service to return true for documentation.enabled
     sinonSandbox
@@ -46,13 +36,13 @@ describe('GET /api/get-documentation', () => {
 
     // stub the documentation service's file suffixes, as the specs are accessing ts files
     sinonSandbox
-      .stub(DocumentationService, 'controllerFileSuffix')
+      .stub(DocumentGeneratorUtil, 'controllerFileSuffix')
       .value('-controller.ts');
     sinonSandbox
-      .stub(DocumentationService, 'routerFileSuffix')
+      .stub(DocumentGeneratorUtil, 'routerFileSuffix')
       .value('-router.ts');
     sinonSandbox
-      .stub(DocumentationService, 'serializerFileSuffix')
+      .stub(DocumentGeneratorUtil, 'serializerFileSuffix')
       .value('-serializer.ts');
 
     const res = await chai
@@ -65,6 +55,23 @@ describe('GET /api/get-documentation', () => {
     expect(res.body.markdownDocumentation).to.be.eq(sampleDocumentation);
   });
 
+  it('should return an empty string, as the documentation is disabled for the testing environment', async () => {
+    // stub the config service to return false for documentation.enabled
+    sinonSandbox
+      .stub(ConfigService, 'getValue')
+      .withArgs('documentation.enabled')
+      .returns(false);
+
+    const res = await chai
+      .request(app)
+      .get('/api/get-documentation')
+      .set('content-type', 'application/json');
+
+    expect(res.status).to.be.eq(200);
+    expect(openAiStub.calledOnce).to.be.false;
+    expect(res.body.markdownDocumentation).to.be.eq('');
+  });
+
   it('should throw an error when it is unable to access the required files', async () => {
     // stub the config service to return true for documentation.enabled
     sinonSandbox
@@ -74,7 +81,7 @@ describe('GET /api/get-documentation', () => {
 
     const nonExistingJsRouteFilePath = `${
       DocumentationService.expressRoutesList[0].rootFolderPath
-    }/${DocumentationService.restApiFolderPath}`;
+    }/rest-api`;
 
     const res = await chai
       .request(app)
